requirements.txt
duckdb==0.10.0
polars==0.20.2
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
plotly==5.18.0
streamlit==1.31.0
pyyaml==6.0.1
shapely==2.0.2
duckdb
PyYAML
pandas
scikit-learn
plotly